+++ back/src/schemas/users/userPatchSchema.ts
@@ -3,9 +3,17 @@ import { userSchema } from './userSchema'
 
 const basePatchSchema = userSchema.pick({ id: true })
 const optionalUserSchema = userSchema
-  .omit({ id: true, avatarId: true, specialization: true })
+  .omit({
+    id: true,
+    avatarId: true,
+    specialization: true,
+    updatedAt: true,
+    createdAt: true,
+  })
   .extend({ specializationId: z.string().cuid() })
   .partial()
 export const userPatchSchema = basePatchSchema
   .merge(optionalUserSchema)
   .strict()
+
+export type UserPatchSchema = z.infer<typeof userPatchSchema>