+++ front/src/styles/FlexBox.ts
@@ -13,7 +13,10 @@ type TFlexBox = {
   gap?: string
 }
 
-export const FlexBox = styled.div<TFlexBox>`
+export const FlexBox = styled.div.withConfig({
+  shouldForwardProp: (prop) =>
+    !['direction', 'justify', 'align', 'gap'].includes(prop),
+})<TFlexBox>`
   display: flex;
   flex-direction: ${({ direction }) => direction || 'column'};
   justify-content: ${({ justify }) => justify || 'center'};