+++ packages/ui/src/__tests__/atoms/Icon.test.tsx
@@ -17,7 +17,7 @@ describe('Icon', () => {
     expect(icon).toHaveClass('material-symbols-outlined')
 
     expect(icon).toHaveStyle(
-      `font-variation-settings: 'FILL' 1,'wght' 400,'GRAD' 0,'opsz' 48;`
+      `font-variation-settings: 'FILL' 1, 'wght' 400,'GRAD' 0, 'opsz' 48;`
     )
     expect(icon).toHaveStyle(`color: ${colors.primary}`)
   })