+++ packages/ui/src/components/atoms/index.ts
@@ -1,12 +1,26 @@
 import Input from './Input'
 import Label from './Label'
-import Icon from './Icon'
+import Checkbox from './Checkbox'
 import { Button } from './Button'
 import HamburgerMenu from './HamburgerMenu'
 import ValidationMessage from './ValidationMessage'
 import Radio from './Radio'
 import Text from './Text'
+import { Spinner } from './Spinner'
+import Counter from './Counter'
 
+export { Avatar } from './Avatar'
+export { Icon } from './Icon'
 export { Title } from './Title'
-export { Input, Label, Button, HamburgerMenu, Icon, ValidationMessage, Text, Radio }
-
+export {
+  Input,
+  Label,
+  Button,
+  Checkbox,
+  HamburgerMenu,
+  ValidationMessage,
+  Text,
+  Radio,
+  Spinner,
+  Counter,
+}