+++ packages/ui/src/components/molecules/index.ts
@@ -1,5 +1,6 @@
 import { BackButton } from './BackButton'
-import InputGroup from './InputGroup'
+import { Tabs } from './Tabs'
 
+export { InputGroup } from './InputGroup'
 export { Modal } from './Modal'
-export { BackButton, InputGroup }
+export { BackButton, Tabs }